Make sure connections are associated to the correct site
If a site is added or removed, loop over current connections and make
sure to associate them to the correct site. This is done by forcibly
re-acquiring them and releasing the previously acquired one.

If not done, the maximum number of connections may temporarily be
exceeded: connections of a running download were not accounted for when
a new site (matching the download) was added.