Better handling of new connections attempts
Resuming a download now only changes its state to 'pending' without
trying a new connection for this download. The download manager is
responsible to try initiating a new connection, and does so for all
downloads in listed order (list considered to be in priority order).
When the attempt is done for a given download, the manager goes to
the next download to allow it to try initiating a new connection,
and so on.
Attempt is considered done in the following cases:
  - a new segment was not started (e.g. limit reached)
  - the connection attempt failed
  - the connection attempt was not completed before a given time:
    instead of the usual cnx timeout (which may be large), it is
    the delay to wait after a cnx error that is used, which also
    makes sense since upon failure we consider the attempt done
    (even though a new one may be done after this delay)

This gives a better behaviour when starting/resuming more than one
download (e.g. upon restarting the application) since in this case
the new connection will be given to the first download in priority
etc., while previously all the concerned downloads were competing
to acquire new connections, resulting in the first 2-3 to be able
to gain one (while technically the first download could monopolize
all granted connections).
When applicable the cases of failure and timeout helps granting
connections to the next download(s) in a reasonable time.