Better download rate computation, with less JavaFX updating
Adjust rate limiter time slice start to a predictable value, aligned
with the rate computation which can then more finely ignore the current
slice.
Rate computation is done more often, but not in the JavaFX thread,
and rate (+ETA) displaying is done a bit less often. Computation is
also nearer to live value due to a smaller time slice (aligned with
the rate limiter).
Rate computation is also more accurate when there is no rate limiting.

Download progress precision is limited to 1/1000 so that UI does not
need to be updated for small changes (that cannot even be visually
seen).

Minor CSS cleanup.