Handle server accepting ranges for download with unknown size
Even though size is unknown, if server accept ranges we can at least
resume the download starting from what was already downloaded.
This implies that we also save the 'downloaded' size and can restore
it properly to resume the download.

Also cleaned/simplified some code dealing with the download size.