When connection cannot be acquired, fail the connection attempt
When the download manager tries to initiate a new connection, if a given
download fails to acquire a new connection (that is, before even trying
to do a request), fail the associated promise so that the attempt can be
done on the next download if any.
Otherwise the promise is not properly completed and only the timeout
would make the manager move to the next download.