Better handle multiple concurrent connection attempts
A download can only start more connections/segments after its very first
connection has started.
If the download manager tries to initiate a new connection at the same
time, the download cannot honor it right away. In this case, don't fail
the attempt (and thus don't go to the next download right away) but link
it with the one associated to the initial connection attempt. Thus if it
succeeds in a timely manner (a timeout is running), this download can
effectively grab all the connections it can use, instead of having those
connections immediately available to other downloads.