Filter unwanted files from jar, and handle devel logback conf
Move resources that are not needed in the generated jars to a
dedicated folder (used upon 'install' task).

Handle a dedicated 'package' resources subfolder as content to
package over existing files: content is put in the generated jars
with relative hierarchy, possibly replacing existing content.
e.g. the 'resources/package/logback.xml' file replaces the
'resources/logback.xml' which is bundled as default resource
file, allowing to have a local devel logback configuration that
won't be bundled in generated jar (replaced by a more generic and
dedicated configuration).

Also filter files that can be generated when running the application
from the IDE. Not necessary when packaging from a clean state, but
useful upon developing.