Prefer all running segments over only started ones
Most of the time (to determine and display ongoing segments) we
actually want to consider all the known segments, whether they are
started or not, and not only the active (started) ones.

Rework the way we update download state icon and segments counts,
as well as downloads counts. Now update them separately by only
listening to concerned properties; this prevents updating too
often some of them 'for nothing', especially the downloads counts.