Better handling of forced connections
Current handling was properly working when forcing new connections to go
beyond the max segments limit. But if used to circumvent the number of
connections limit (e.g. to force a download to start), when a forced
connection would end (with issue or not), it would not automatically
trigger a new one.
The new handling remember the number of forced connections that were
created (and thus are allowed to exist) and works in either situation.