Unversion the application JAR file upon installing
Keeping the version in the filename is not needed; it is present
in the main window title.

The script to launch the app can point to the unversioned name.
Also add the 'lib' folder to the classpath: we can drop additional
JARs if needed (e.g. javafx-swing so that scenic-view works), or
handle a nominal application JAR (not assembly made) with its
dependencies inside the 'lib' folder.