Handle exiting application when classes are missing
At least on Windows: if the JAR is overwritten while the application
is running, access to classes not yet loaded by the JVM will trigger
NoClassDefFoundError.
It especially happens when user requests to exit the application. In
this specific case, print the issue and trigger nominal JVM exiting
(without trying to properly end running downloads since it is now
next to impossible).