Keep new download window open after user requested renaming file
When a path is already taken, the new download dialog asks whether
to resume/restart/rename. Upon renaming, select a new available
path and let the dialog open with the updated path. This allows the
user to change the proposed name, instead of simply notifying of
the new name and close the dialog (without possible interaction).