Prefer IPv6 when client and server handle both IPv4 and IPv6
Needed when intercepting some downloads for which a temporary URL
has been built for the client IP: usually the browser will prefer
the IPv6, so if dl-mngr does use the IPv4 one, server will fail
the request.

Since the behaviour is determined by an environment variable which
is read only once at VM startup and then cached, it must be set in
the JVM command-line parameters.