Handle optional pattern to match site hosts
The nominal behaviour is to check the target download host against
site names to determine whether it belongs to one configured site.
If it does not match, the parent host name is recursively used to
check again, until either a site matches or we reached a Top-Level
Domain (which is not checked).
e.g. host 'a.b.c' can belong to site 'a.b.c', or 'b.c'.
Note: host name is lower-cased before checking.

For each site, a pattern can now be added.
The full host is still checked first for an exact site match. But
if it does not match, before walking up its parent, the full host
is now checked against all site patterns (in no particular order).

If a match is found for a pattern, the host belongs to the site.
e.g.: 'site\.name' matches 'site.name', 'child.site.name' but also
'site.name.parent'.
Partial or full matching is possible with the appropriate pattern.
e.g.: 'site\.name$', '^my\.full\.site\.name$', ...