Disable all site option fields when no site is selected
If no site is selected, one could think that filling the site name
field (and other options) before hitting the 'Add' button would add
a new site with the filled settings.
This is not the case.

Since it may be troublesome to try to handle it, and there is not
much advantage over hitting 'Add' then filling the site settings,
disable all site fields in this case in order to leave no room for
misunderstanding.
Also don't toggle error on the site name field which is then empty
for reason in this case.