Make sure main window remain iconified after adding new download
Works natively on Windows (10).
But on Gnome (3.34) the main (parent) window is de-iconified when
dealing with a modal dialog (either when showing the dialog, or
when closing it).
Before showing the modal dialog, check whether the main stage is
iconified, and re-iconify it upon closing the dialog: this does
nothing on Windows (stage had remained iconified - unless user
did bring it back), but does the trick on Gnome.