Complete promise before building dialog to add new download
We notify caller when the main controller did take into account
the new download, not whether we fail, user discard it or when
it actually is added.
Do it before building the dialog to add it, so that caller is
still properly notified even if we fail to build the dialog.
Upon issue, the error is displayed in the user interface and
in the logs, which is what we need.