Better handle new download path sanitization
Refrain from using Paths.get on non-sanitized values, as it fails
when given 'illegal characters' (subset of reserved characters we
do want to sanitize).
As such, sanitize the requested - through CLI or WebSocket - new
download filename before using it.

Not only sanitize the filename, but also the folder path: user may
manually edit the field in the dialog.