Handle unrecoverable error after completed download
If file was fully downloaded but an error occurred when finalizing it
(e.g. renaming from temporary to final name), we must not try to resume
the download as there is nothing more to do: with previous code it
remained in 'pending' state and could not be properly resumed/stopped
possibly preventing the application to properly end.

Immediately finish download if trying to resume one which has no segment
remaining.

Prevent inner error when target file does not exist anymore while we are
finalizing the download: would end the running actor, triggering bad
side effects. Instead trigger an explicit, and handled, error which,
combined with the other changes, lets the download in the failed state.

Fix one code warning.