Make sure to not (re-)create site entry in settings
The function to get a known site entry is also used to create it if
not existing. In most cases there is no issue with it, as caller
does request a site that is expected to be known (often it was
determined right before).

But there is at least one situation where this is not the case:
 - a site is actively cached in the download manager
 - this site is removed through the options controller
Then the download manager is requested to refresh itself, which as
a side effect does 'get' (and re-creates) the site entry we just
removed.

To fix this we first drop unknown (and thus removed) sites from
the download manager cache while refreshing it.
Then the function to get a site entry now fallbacks to the default
site if requested one is unknown, while warning about this access.
And a new function to create a new site entry in the settings is
reserved for the case where we actually need it.