Disable 'quiet' netty period when shutting down WebSocket
It allows activity to continue, and shutdown to be delayed, as long
as new tasks are submitted whithin the given period - with a timer
that is reset each time a task is submitted.
We want to shutdown ASAP, so we don't need it.
Also reduce the shutdown timeout from 15s (netty's default) to 10s.