Handle JSON through CLI
When requested ('--json' CLI option), CLI parameters are read from
stdin. They are merged in the ones received thourh the CLI command.
A one-line JSON object is expected, with the same format than through
the WebSocket connection.
The CLI response is also done as JSON (one-line too).

This has a number of benefits:
 - the first unique instance started would only have the '--json' CLI
   option visible in process list
 - more complex parameters can be passed as JSON compared to what is
   possible with CLI option parsing
 - bigger data can be passed through stdin compared to CLI
 - passing JSON through CLI allows easier parity with WebSocket usage
 - command output would be able to contain more information if needed