cleanup: remove unneeded parameter
In most cases, reuseOpt is None.
reuseOpt is only filled when adding a download in the list.

When the download is brand new, 'reuse' is already properly set so
reuseOpt is simply redundant.

When a pre-existing download is found, 'reuseOpt' forces 'reuse' to
true, which is actually wrong.
If the download had started, this does not change anything as 'reuse' is
already true (file created).
If the download did not start yet, 'reuse' is false and forcing it to
true allows truly reusing a pre-existing temporary file: this is wrong
because if file exists, it was not created by this download.