fix: validate downloaded content sooner
The download manager listens to download completion promise, which is
asynchronous relatively to the actual file download actor.
Yet the actor needs to know whether the downloaded content is valid or
not in order to properly stop (or remain active).

Move code where download is directly completed, so that content is
checked when appropriate, before the result is used to determine what to
do.