hls: pre-create target, and remember it
If we need to resume HLS processing, and target was created in a
previous attempt, code would see the name is not available and select an
alternate name " (n)" suffix. Then again each time we resume.
Instead, the first time we process HLS, pre-create (to reserve) an
available target name, and remember that we did it, to reuse it later.